Created a specific config for watson-extension
