+++ libs/common/src/common/config/__init__.py
@@ -24,5 +24,45 @@ __repository_chain.append(RepositoryOpenshift())
 
 config = Config(ChainMap(*__repository_chain))
 
+def log_config(module):
+    import logging
+
+    primitives = (bool, str, int, float, type(None))
+
+    def is_primitive(obj):
+        return isinstance(obj, primitives)
+
+    def should_log(key, value) -> bool:
+        if not is_primitive(value):
+            return False
+
+        if key.startswith("_"):
+            return False
+
+        return True
+
+    def get_value(key: str, value) -> str:
+        if value is None or (
+                isinstance(value, int) and not isinstance(value, bool) and value == 0
+        ):
+            return f"--not-set-- ({value})"
+
+        upper_key = key.upper()
+        accepted_variables = ["dev_sso_refresh_token_url"]
+        if (
+                any(
+                    banned in upper_key for banned in ["PASSWORD", "TOKEN", "SECRET", "KEY"]
+                )
+                and key not in accepted_variables
+        ):
+            return "*********"
+
+        return str(value)
+
+    for k, v in module.__dict__.items():
+        if should_log(k, v):
+            logging.info(f"Using {k}: {get_value(k, v)}")
+
+
 # We could implement expanvars If we need variable expansion
 # https://github.com/sayanarijit/expandvars
